OrderGroup create API 정상작동

+++ study/src/main/java/com/example/study/model/entity/OrderGroup.java
@@ -40,7 +40,7 @@ public class OrderGroup {
 
     private BigDecimal totalPrice;
 
-    private Integer totalQuantity;
+    private Long totalQuantity;
 
     private LocalDateTime orderAt;
